[REFACTOR] 공통 Header를 사용하도록 변경
