+++ src/pages/itemDetail/ItemDetailPage.tsx
@@ -1,8 +1,8 @@
 import ConfirmModal from "@/components/common/ConfirmModal";
+import Header from "@/components/common/Header";
 import BidBottomSheet from "@/components/item/detail/bid/BidBottomSheet";
 import BidHistoryList from "@/components/item/detail/bidHistory/BidHistoryList";
 import Footer from "@/components/item/detail/footer/Footer";
-import Header from "@/components/item/detail/header/Header";
 import ItemCard from "@/components/item/detail/itemCard/ItemCard";
 import ItemImage from "@/components/item/detail/itemImage/ItemImage";
 import QnaInputBar from "@/components/item/detail/qna/QnaInputBar";
